// generated/backend/InitializedProfile.ts
/**
 * Do not edit this file it is auto-generated by italia-utils / gen-api-models.
 * See https://github.com/teamdigitale/italia-utils
 */
/* tslint:disable */

import { EmailAddress } from "./EmailAddress";
import { BlockedInboxOrChannels } from "./BlockedInboxOrChannels";
import { PreferredLanguages } from "./PreferredLanguages";
import { IsInboxEnabled } from "./IsInboxEnabled";
import { IsWebhookEnabled } from "./IsWebhookEnabled";
import { FiscalCode } from "./FiscalCode";
import { HasProfile } from "./HasProfile";
import { Version } from "./Version";
import * as t from "io-ts";
import { NonEmptyString } from "italia-ts-commons/lib/strings";

/**
 * Describes the user's profile after it has been stored in the Profile API.
 */

// required attributes
const InitializedProfileR = t.interface({
  is_inbox_enabled: IsInboxEnabled,

  is_webhook_enabled: IsWebhookEnabled,

  family_name: t.string,

  fiscal_code: FiscalCode,

  has_profile: HasProfile,

  name: t.string,

  spid_email: EmailAddress,

  spid_mobile_phone: NonEmptyString,

  version: Version
});

// optional attributes
const InitializedProfileO = t.partial({
  email: EmailAddress,

  blocked_inbox_or_channels: BlockedInboxOrChannels,

  preferred_languages: PreferredLanguages
});

export const InitializedProfile = t.exact(
  t.intersection(
    [InitializedProfileR, InitializedProfileO],
    "InitializedProfile"
  )
);

export type InitializedProfile = t.TypeOf<typeof InitializedProfile>;